Update to use call-specific #errno value.
Added a FsLibraryGenerator which generated calls to library functions which support #errno. Instead of taking all of the arguments as individual parameters, this generator expects them to be passed as an Array.

Previously you might have used a generated library in this way.
```smalltalk
{ #category : 'Functions' }
_libcFcntl >> open_: __file _: __oflag [

	^Functionopen callWith: { __file. __oflag"varArgs array should be appended to this array" }

]

{ #category : 'Functions' }
_libcFcntl >> open_: __file _: __oflag _: __varArgs [

  ^ Functionopen
		callWith:
			{__file.
			__oflag} , __varArgs
]
```

Now, one function is generated.

```smalltalk
open: args errno: errnoHolder

	^Functionopen
		callWith: args
		errno: errnoHolder
```

You call the function by generating an array with the arguments.

```smalltalk
errnoHolder := { 0 }.
result := library
  open: { file. flags. }
  errno: errnoHolder.
```

This pattern is much easier on the eyes. It also avoids the need to generate a varArgs array separate from the normal arguments.