Migrated to using FileOpeningOptions
When opening a file, we now create a FileOpeningOptions classes instead of specifying whether a file is read
able or writable via boolean flag. This will allow us to allow both reading and writing for the same descrip
tor. It will also removes hard-coded values passed to the FFI. The values such as mode are calculated based
upon how the UnixFileOpeningOptions is configured.

Added tests for MemoryFileOpeningOptions.