50875: Ensure symlinks to files or directories are included in the result of #files or #directories.
As part of 50730, when the type of entity was unknown while enumerating directory entries, `lstat` would be invoked to determine the type of entity. Per 50875, symlinks to files or directories should be included in the result to #files or #directories.

Replacing the `lstat` call with a `stat` call ensures that we behave based on the type of the target rather than the local directory entry type.