140: Initial wrapper functions for POSIX functions.
This variant of the wrapper functions does not attempt to change the behavior in the image. Function calls take the same arguments. Issues discovered during review have not been fixed. For instance, getcwd will still be called without allocating a buffer -- which is undefined according to POSIX.

Further revision is required but this commit will permit the backend code to remove OS/Platform concerns.